requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-appconf==1.0.3
django-crispy-forms==1.8.0
django-grappelli==2.13.1
django-imagekit==4.0.2
django-ok-likes==0.4
djangorestframework==3.10.3
idna==2.8
mysqlclient==1.4.4
pilkit==2.0
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.3
